Rename chpl_machine -> chpl_arch
