+++ util/chplenv/chpl_bin_subdir.py
@@ -4,14 +4,14 @@ import optparse
 import sys
 
 import chpl_home_utils
-import chpl_platform, chpl_machine, chpl_cpu
+import chpl_platform, chpl_arch, chpl_cpu
 
 from utils import memoize
 
 @memoize
 def get(flag='host'):
     platform = chpl_platform.get(flag)
-    machine = chpl_machine.get(flag)
+    machine = chpl_arch.get(flag)
     (_, arch) = chpl_cpu.get(flag, map_to_compiler=True,
                              get_lcd=chpl_home_utils.using_chapel_module())
