+++ util/chplenv/chplenv.py
@@ -14,7 +14,7 @@ import chpl_jemalloc
 import chpl_launcher
 import chpl_llvm
 import chpl_locale_model
-import chpl_machine
+import chpl_arch
 import chpl_mem
 import chpl_regexp
 import chpl_tasks
