+++ util/chplenv/printchplenv.py
@@ -135,11 +135,11 @@ def compute_all_values():
     ENV_VALS["CHPL_ORIG_TARGET_COMPILER"] = chpl_compiler.get('target', llvm_mode='orig')
     ENV_VALS['CHPL_HOST_PLATFORM'] = chpl_platform.get('host')
     ENV_VALS['CHPL_HOST_COMPILER'] = chpl_compiler.get('host')
-    ENV_VALS['CHPL_HOST_MACHINE'] = chpl_machine.get('host')
+    ENV_VALS['CHPL_HOST_MACHINE'] = chpl_arch.get('host')
     ENV_VALS['CHPL_HOST_CPU'] = chpl_cpu.get('host').arch
     ENV_VALS['CHPL_TARGET_PLATFORM'] = chpl_platform.get('target')
     ENV_VALS['CHPL_TARGET_COMPILER'] = chpl_compiler.get('target')
-    ENV_VALS['CHPL_TARGET_MACHINE'] = chpl_machine.get('target')
+    ENV_VALS['CHPL_TARGET_MACHINE'] = chpl_arch.get('target')
     ENV_VALS['CHPL_TARGET_CPU'] = chpl_cpu.get('target').arch
 
     # Use module's LCD architecture in case it was built before
